Add locking for database updating
Due to potential issues with database locking (blocking), SysV
semaphores (blocking, at least with PHP's restrictions), and flock
advisory locking (doesn't work on network filesystems, which
classroom.cs.unc.edu uses), the locking method is fopen(..., "x") which
has the behavior atomically creating a file, or failing because the file
exists.

Currently, the lock is acquired only when responding to a POST /update,
but if the lock isn't acquired then the request succeeds without doing
anything. The idea is to allow multiple clients to POST /update if
asked, and let only the first one actually do any work.